+++ gss-domain/src/test/java/com/devoops/domain/entity/analysis/OpenAiModelTest.java
@@ -3,6 +3,7 @@ package com.devoops.domain.entity.analysis;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
+import java.math.BigDecimal;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
@@ -13,8 +14,8 @@ class OpenAiModelTest {
 
         @Test
         void GPT5_범위에서_선택된다() {
-            int lowerBound = OpenAiModel.GPT_5.getMoneyUnderCriteria();
-            int upperBound = OpenAiModel.GPT_5.getMoneyUpperCriteria();
+            BigDecimal lowerBound = BigDecimal.valueOf(OpenAiModel.GPT_5.getMoneyUnderCriteria());
+            BigDecimal upperBound = BigDecimal.valueOf(OpenAiModel.GPT_5.getMoneyUpperCriteria());
 
             assertAll(
                     () -> assertThat(OpenAiModel.getModelByUsage(lowerBound)).isEqualTo(OpenAiModel.GPT_5),
@@ -24,8 +25,8 @@ class OpenAiModelTest {
 
         @Test
         void GPT5_MINI_범위에서_선택된다() {
-            int lowerBound = OpenAiModel.GPT_5_MINI.getMoneyUnderCriteria();
-            int upperBound = OpenAiModel.GPT_5_MINI.getMoneyUpperCriteria();
+            BigDecimal lowerBound = BigDecimal.valueOf(OpenAiModel.GPT_5_MINI.getMoneyUnderCriteria());
+            BigDecimal upperBound = BigDecimal.valueOf(OpenAiModel.GPT_5_MINI.getMoneyUpperCriteria());
 
             assertAll(
                     () -> assertThat(OpenAiModel.getModelByUsage(lowerBound)).isEqualTo(OpenAiModel.GPT_5_MINI),
@@ -35,8 +36,8 @@ class OpenAiModelTest {
 
         @Test
         void GPT5_NANO_범위에서_선택된다() {
-            int lowerBound = OpenAiModel.GPT_5_NANO.getMoneyUnderCriteria();
-            int upperBound = OpenAiModel.GPT_5_NANO.getMoneyUpperCriteria();
+            BigDecimal lowerBound = BigDecimal.valueOf(OpenAiModel.GPT_5_NANO.getMoneyUnderCriteria());
+            BigDecimal upperBound = BigDecimal.valueOf(OpenAiModel.GPT_5_NANO.getMoneyUpperCriteria());
 
             assertAll(
                     () -> assertThat(OpenAiModel.getModelByUsage(lowerBound)).isEqualTo(OpenAiModel.GPT_5_NANO),
